+++ website/src/examples/dynamic-row-heights.js
@@ -208,7 +208,6 @@ export default class App extends React.Component {
         >
           <Table
             fixed
-            selectable
             columns={
               !!this.state.toggle ? this.columns.slice(0, 4) : this.columns
             }